+++ CHANGELOG.md
@@ -29,6 +29,10 @@ This project adheres to [Semantic Versioning](https://semver.org).
 * ``MobileClient.playlist_entries``.
 * ``MobileClient.playlist_entries_iter``.
 
+### Fixed
+
+* Token refreshing when providing token argument to client classes.
+
 
 ## [2.1.0](https://github.com/thebigmunch/google-music/releases/tag/2.1.0) (2018-11-26)